requirements.txt
streamlit
google-generativeai
python-dotenv
scikit-learn
nltk

